Update thread detection to not include mentions
Noticed 're:' was being added to notes in the UI when it only mentioned
a note. This didn't constitute a thread and shouldn't attempt to add who
it is replying to in the UI since there wasn't a reply to display. This
simply makes sure that a thread is detected when there is an 'e' tag and
the marker for the tag is not a mention.